Add sddf as a submodule
The sdfgen tool/library does not depend on sDDF at build time,
so having this as a submodule is just for testing and
experimentation purposes.

Eventually we can just have it as a Zig build dependency, but I
don't want to do that while there are a lot of changes in sDDF
(e.g the configuration files that the sdfgen tool depends on)
ongoing.