Progress on making pysdfgen
This is a Python module so that people can generate SDF files from
Python because most people probably don't want to use straight Zig,
which is fair enough.

Because of Zig's interoperability with C, doing this is fairly
straight-forward, meaning that connecting to other languages such as
Rust or Go shouldn't be too bad either.