Adding abstractions for serial sDDF sub-system
This patch adds functionality for creating a serial
sub-system that matches the sDDF architecture and protocols
from a high-level.

Originally I wanted this to be done more from configuration
files but struggled to find a correct abstraction that would
fit across multiple device classes. It is especially hard
given that sDDF only works on a few device classes.

For now, we have decided to do this purely programmatically,
with only the device driver specific parts being in a config
file in sDDF.

Over time, we will see what abstractions work and don't and
can possibly find a suitable configuration file format that
expresses what we need to auto-generate SDF files for sDDF
systems.