Use @panic in unrecoverable errors.
Right now this just simplifies a lot of the API that could fail
due to OOM. If that happens there's nothing really we can do and
neither can the user in this contetx I think, so we just panic.

This makes the API simpler since people don't have to handle the
error now.